# api-gateway/internal/app/docs/swagger.yaml
basePath: /
definitions:
  bookingentity.CreateBookingReq:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
      total_amount:
        type: string
    type: object
  entity.AddRoomToHotelsReq:
    properties:
      availability:
        type: string
      price_per_night:
        type: string
      room_type:
        type: string
    type: object
  entity.BookingRoomReq:
    properties:
      book:
        type: boolean
    type: object
  entity.CreateHotelServiceReq:
    properties:
      address:
        type: string
      location:
        type: string
      name:
        type: string
      rating:
        type: string
    type: object
...
  entity.GetAllHotelRes:
    properties:
      count:
        type: integer
      hotels:
        items:
          $ref: '#/definitions/entity.Hotel'
        type: array
    type: object
...
  entity.GetHotelsRoomsRes:
    properties:
      count:
        type: integer
      rooms:
        items:
          $ref: '#/definitions/entity.Room'
        type: array
    type: object
  entity.Hotel:
    properties:
      address:
        type: string
      hotel_id:
        type: string
      location:
        type: string
      name:
        type: string
      rating:
        type: string
    type: object
...
  entity.Room:
    properties:
      availability:
        type: string
      price_per_night:
        type: string
      room_id:
        type: string
      room_is_busy:
        type: boolean
      room_type:
        type: string
    type: object
...
  entity.UpdateHotelReq:
    properties:
      address:
        type: string
      hotel_id:
        type: string
      location:
        type: string
      name:
        type: string
      rating:
        type: string
    type: object
  entity.UpdateHotelsRoomReq:
    properties:
      availability:
        type: string
      price_per_night:
        type: string
      room_type:
        type: string
    type: object
...
host: localhost:9006
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Artisan Connect
  version: "1.0"
paths:
  /booking/add/waitlist/{hotel_id}/{room_type}:
    post:
      consumes:
      - application/json
      description: AddUserToWaitList hotel
      parameters:
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      - description: room_type
        in: path
        name: room_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: AddUserToWaitList
      tags:
      - booking
  /booking/create/{hotel_id}/{room_type}:
    post:
      consumes:
      - application/json
      description: CreateBooking hotel
      parameters:
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      - description: room_type
        in: path
        name: room_type
        required: true
        type: string
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bookingentity.CreateBookingReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: CreateBooking
      tags:
      - booking
  /hotel/create:
    post:
      consumes:
      - application/json
      description: Register a new Hotel
      parameters:
      - description: User registration information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateHotelServiceReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Hotel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: CreateHotel
      tags:
      - hotel
  /hotel/delete/{hotel_id}:
    delete:
      consumes:
      - application/json
      description: DeleteHotel hotel
      parameters:
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: DeleteHotel
      tags:
      - hotel
  /hotel/get:
    get:
      consumes:
      - application/json
      description: getting hotel
      parameters:
      - description: Field
        in: header
        name: field
        type: string
      - description: Value
        in: header
        name: value
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.GetAllHotelRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: GetHotels
      tags:
      - hotel
  /hotel/room/book/{hotel_id}/{room_id}:
    put:
      consumes:
      - application/json
      description: UpdateStatusBook hotel
      parameters:
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      - description: room_id
        in: path
        name: room_id
        required: true
        type: string
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.BookingRoomReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Room'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: UpdateStatusBook
      tags:
      - room
  /hotel/room/create/{hotel_id}:
    post:
      consumes:
      - application/json
      description: AddRoomToHotels hotel
      parameters:
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.AddRoomToHotelsReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Room'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: AddRoomToHotels
      tags:
      - room
  /hotel/room/delete/{hotel_id}/{room_id}:
    delete:
      consumes:
      - application/json
      description: DeleteHotelsRooms hotel
      parameters:
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      - description: room_id
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: DeleteHotelsRooms
      tags:
      - room
  /hotel/room/gets/{hotel_id}:
    get:
      consumes:
      - application/json
      description: GetHotelsRooms hotel
      parameters:
      - description: Field
        in: header
        name: field
        type: string
      - description: Value
        in: header
        name: value
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.GetHotelsRoomsRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: GetHotelsRooms
      tags:
      - room
  /hotel/room/update/{hotel_id}/{room_id}:
    patch:
      consumes:
      - application/json
      description: UpdateHotelsRooms hotel
      parameters:
      - description: hotel_id
        in: path
        name: hotel_id
        required: true
        type: string
      - description: room_id
        in: path
        name: room_id
        required: true
        type: string
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateHotelsRoomReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Room'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: UpdateHotelsRooms
      tags:
      - room
  /hotel/update:
    patch:
      consumes:
      - application/json
      description: UpdateHotel hotel
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateHotelReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Hotel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: UpdateHotel
      tags:
      - hotel
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"